requirements.txt
moviepy<2.0
flask
pillow
google-generativeai
python-dotenv
openai
